Improved encapsulation and added unit tests
Added a couple of unit tests, additional unit tests can be added. This will make testing a lot easier.

Separated loading the data from LondonUnderground and moved it to new file LoadMapData

Added Route and RouteFinder classes to extract the ShortestPathAlgorithm into. This needs further refining. The idea, currently the Station is heavily encapsulated with the route, meaning we need to reset the Station. These two should be separated, we can create a RouteNode to contain the temporary Route calculation data.

This still maintains all the previous functionality but it is organised better. The implementation is not finished but this can be a good base for providing better encapsulation. In my opinion, this needs to be refactored more.